pyproject.toml
[build-system]
requires = ["setuptools>=75.3.0"]
build-backend = "setuptools.build_meta"

[project]
name = "auto-recon"
version = "1.0"
authors = [{ name = "Ivan Sincek" }]
description = "Not another auto-reconnaissance framework."
readme = "README.md"
requires-python = ">=3.10"
classifiers = [
	"Programming Language :: Python :: 3",
	"License :: OSI Approved :: MIT License",
	"Operating System :: POSIX :: Linux"
]
dependencies = [
	"bot_safe_agents>=1.0",
	"colorama>=0.4.6",
 	"cryptography>=44.0.1",
	"jq>=1.6.0",
	"pyOpenSSL>24.0.0",
	"regex>=2023.8.8",
	"tabulate>=0.9.0",
	"termcolor>=2.4.0",
	"tldextract>=3.6.0"
]

[project.urls]
"Homepage" = "https://github.com/ivan-sincek/auto-recon"

[project.scripts]
auto-recon = "auto_recon.main:main"

[tool.setuptools]
license-files = []

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
"*" = ["wordlists/*.txt"]
